VP-1677 Separate AccountOrders.vue to vue and ts files
