+++ client-app/src/pages/account/router/index.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import VueRouter from "vue-router";
 import AccountInfo from "@account/views/AccountInfo.vue";
-import AccountOrders from "@account/views/AccountOrders.vue";
+import AccountOrders from "@account/views/account-orders/index.vue";
 import AccountUsers from "@account/views/account-users/index.vue";
 
 Vue.use(VueRouter);
